fix: rename shared package to servicespec
