+++ pkg/cluster/services/kafka.go
@@ -7,7 +7,7 @@ import (
 	"github.com/redhat-developer/app-services-cli/pkg/cluster/v1alpha"
 	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
 	"github.com/redhat-developer/app-services-cli/pkg/kafkautil"
-	"github.com/redhat-developer/app-services-cli/pkg/shared"
+	"github.com/redhat-developer/app-services-cli/pkg/servicespec"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -75,7 +75,7 @@ func (s KafkaService) BuildServiceDetails(serviceName string, namespace string, 
 		Name:               serviceName,
 		KubernetesResource: kafkaConnectionCR,
 		GroupMetadata:      resources.AKCResource,
-		Type:               shared.KafkaServiceName,
+		Type:               servicespec.KafkaServiceName,
 	}
 
 	return &serviceDetails, nil
