+++ pkg/cmd/cluster/bind/bind.go
@@ -12,7 +12,7 @@ import (
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/iostreams"
 	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
 	"github.com/redhat-developer/app-services-cli/pkg/core/logging"
-	"github.com/redhat-developer/app-services-cli/pkg/shared"
+	"github.com/redhat-developer/app-services-cli/pkg/servicespec"
 
 	"github.com/spf13/cobra"
 )
@@ -86,7 +86,7 @@ func NewBindCommand(f *factory.Factory) *cobra.Command {
 	cmd.Flags().StringVar(&opts.serviceName, "service-name", "", opts.localizer.MustLocalize("cluster.common.flag.serviceName.description"))
 
 	_ = cmd.RegisterFlagCompletionFunc("service-type", func(cmd *cobra.Command, _ []string, _ string) ([]string, cobra.ShellCompDirective) {
-		return shared.AllServiceLabels, cobra.ShellCompDirectiveNoSpace
+		return servicespec.AllServiceLabels, cobra.ShellCompDirectiveNoSpace
 	})
 
 	return cmd
