+++ pkg/cmd/status/status.go
@@ -11,8 +11,7 @@ import (
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/iostreams"
 	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
 	"github.com/redhat-developer/app-services-cli/pkg/core/logging"
-
-	"github.com/redhat-developer/app-services-cli/pkg/shared"
+	"github.com/redhat-developer/app-services-cli/pkg/servicespec"
 
 	"github.com/spf13/cobra"
 )
@@ -35,7 +34,7 @@ func NewStatusCommand(f *factory.Factory) *cobra.Command {
 		Config:     f.Config,
 		Connection: f.Connection,
 		Logger:     f.Logger,
-		services:   shared.AllServiceLabels,
+		services:   servicespec.AllServiceLabels,
 		localizer:  f.Localizer,
 		Context:    f.Context,
 	}
@@ -45,12 +44,12 @@ func NewStatusCommand(f *factory.Factory) *cobra.Command {
 		Short:     opts.localizer.MustLocalize("status.cmd.shortDescription"),
 		Long:      opts.localizer.MustLocalize("status.cmd.longDescription"),
 		Example:   opts.localizer.MustLocalize("status.cmd.example"),
-		ValidArgs: shared.AllServiceLabels,
-		Args:      cobra.RangeArgs(0, len(shared.AllServiceLabels)),
+		ValidArgs: servicespec.AllServiceLabels,
+		Args:      cobra.RangeArgs(0, len(servicespec.AllServiceLabels)),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) > 0 {
 				for _, s := range args {
-					if !flagutil.IsValidInput(s, shared.AllServiceLabels...) {
+					if !flagutil.IsValidInput(s, servicespec.AllServiceLabels...) {
 						return opts.localizer.MustLocalizeError("status.error.args.error.unknownServiceError", localize.NewEntry("ServiceName", s))
 					}
 				}
